Issue#1264, Added all views, updated README.md with documentation.
