+++ composite-view/build_artifacts/businessNews.jsp
@@ -1,0 +1,33 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Kevin
+  Date: 11/29/2021
+  Time: 2:51 PM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+    <head>
+        <style>
+            h2 { text-align: center;}
+            table {border: 1px solid black;}
+            tr {text-align: center;}
+            td {text-align: center;}
+        </style>
+    </head>
+    <body>
+        <h2>
+            Generic Business News
+        </h2>
+        <table style="margin-right: auto; margin-left: auto">
+            <tr>
+                <td>Stock prices up across the world</td>
+                <td>New tech companies to invest in</td>
+            </tr>
+            <tr>
+                <td>Industry leaders unveil new project</td>
+                <td>Price fluctuations and what they mean</td>
+            </tr>
+        </table>
+    </body>
+</html>
