+++ composite-view/build_artifacts/index.jsp
@@ -1,0 +1,20 @@
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+  <head>
+    <style>
+      h1 { text-align: center;}
+      h2 { text-align: center;}
+      h3 { text-align: center;}
+    </style>
+  </head>
+  <body>
+    <h1>Welcome To The Composite Patterns Mock News Site</h1>
+    <h2>Send a GET request to the "/news" path to see the composite view with mock news</h2>
+    <h2>Use the following parameters:</h2>
+    <h3>name: string name to be dynamically displayed</h3>
+    <h3>bus: boolean for whether you want to see the mock business news</h3>
+    <h3>world: boolean for whether you want to see the mock world news</h3>
+    <h3>sci: boolean for whether you want to see the mock world news</h3>
+    <h3>sport: boolean for whether you want to see the mock world news</h3>
+  </body>
+</html>
