+++ composite-view/build_artifacts/localNews.jsp
@@ -1,0 +1,25 @@
+
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+    <body>
+        <div style="text-align: center">
+            <h3>
+                Generic Local News
+            </h3>
+            <ul style="list-style-type: none">
+                <li>
+                    Mayoral elections coming up in 2 weeks
+                </li>
+                <li>
+                    New parking meter rates downtown coming tomorrow
+                </li>
+                <li>
+                    Park renovations to finish by the next year
+                </li>
+                <li>
+                    Annual marathon sign ups available online
+                </li>
+            </ul>
+        </div>
+    </body>
+</html>
