+++ composite-view/build_artifacts/scienceNews.jsp
@@ -1,0 +1,34 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Kevin
+  Date: 11/29/2021
+  Time: 4:18 PM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+    <body>
+        <div style="text-align: center">
+            <h3>
+                Generic Science News
+            </h3>
+            <ul>
+                <li>
+                    New model of gravity proposed for dark matter
+                </li>
+                <li>
+                    Genetic modifying technique proved on bacteria
+                </li>
+                <li>
+                    Neurology study maps brain with new precision
+                </li>
+                <li>
+                    Survey of rainforest discovers 15 new species
+                </li>
+                <li>
+                    New signalling pathway for immune system discovered
+                </li>
+            </ul>
+        </div>
+    </body>
+</html>
