+++ composite-view/build_artifacts/sportsNews.jsp
@@ -1,0 +1,32 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Kevin
+  Date: 11/29/2021
+  Time: 3:53 PM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+  <head>
+    <style>
+      h2 { text-align: center;}
+      table {border: 1px solid black;}
+      tr {text-align: center;}
+      td {text-align: center;}
+    </style>
+  </head>
+  <body>
+    <h2>
+      Generic Sports News
+    </h2>
+    <div style="margin-left: auto; margin-right: auto; padding: 20px">
+      International football match delayed due to weather, will be held next week
+    </div>
+    <div style="margin-left: auto; margin-right: auto; padding: 20px">
+      New rising stars in winter sports, ten new athletes that will shake up the scene
+    </div>
+    <div style="margin-left: auto; margin-right: auto; padding: 20px">
+      Biggest upset in basketball history, upstart team sweeps competition
+    </div>
+  </body>
+</html>
