+++ composite-view/build_artifacts/worldNews.jsp
@@ -1,0 +1,34 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Kevin
+  Date: 11/29/2021
+  Time: 2:51 PM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <style>
+        h2 { text-align: center;}
+        table {border: 1px solid black;}
+        tr {text-align: center;}
+        td {text-align: center;}
+    </style>
+</head>
+<body>
+    <h2>
+        Generic World News
+    </h2>
+    <table style="margin-right: auto; margin-left: auto">
+        <tr>
+            <td>New trade talks happening at UN on Thursday</td>
+        </tr>
+        <tr>
+            <td>European Union to announce new resolution next week</td>
+        </tr>
+        <tr>
+            <td>UN delivers report on world economic status</td>
+        </tr>
+    </table>
+</body>
+</html>
